И ещё работа над кешем

+++ classes/bors/cache/static.php
@@ -4,6 +4,7 @@ class cache_static extends base_object_db
 {
 	function main_db() { return config('cache_database'); }
 	function main_table() { return 'cached_files'; }
+	function storage_engine() { return 'bors_storage_mysql'; }
 	function main_table_fields()
 	{
 		return array(
@@ -16,14 +17,36 @@ class cache_static extends base_object_db
 			'target_class_id' => 'class_id',
 			'target_id' => 'object_id',
 			'recreate',
+			'bors_site',
 		);
 	}
 
 	function auto_targets()
 	{
 		return array_merge(parent::auto_targets(), array(
-			'target' => 'target_class_id(target_id)',
+			'_target' => 'target_class_id(target_id)',
 		));
+	}
+
+	function target()
+	{
+		if($this->__havefc())
+			return $this->__lastc();
+
+		$x = NULL;
+		if($this->original_uri())
+			$x = bors_load_uri($this->original_uri());
+
+		if(!$x && $this->object_uri())
+			$x = bors_load_uri($this->object_uri());
+
+		if(!$x)
+		{
+			echo "Can't load {$this->original_uri()} nor {$this->object_uri()}\n";
+			$x = bors_load($this->target_class_name(), $this->target_id());
+		}
+
+		return $x;
 	}
 
 	static function drop($object)
@@ -31,12 +54,18 @@ class cache_static extends base_object_db
 		if(!$object || !config('cache_database'))
 			return;
 
-		$caches = bors_find_all('cache_static', array('target_class_id' => $object->extends_class_id(), 'target_id' => $object->id()));
+		$caches = bors_find_all('cache_static', array(
+			'target_class_id' => $object->extends_class_id(),
+			'target_id' => $object->id(),
+			'by_id' => true,
+		));
 
 		if(file_exists($object->static_file()))
-			if($cache = object_load('cache_static', $object->static_file()))
-				$caches[] = $cache;
+			if($cache = bors_load('cache_static', $object->static_file()))
+				$caches[$cache->id()] = $cache;
 
+		$first = true;
+		$cache = NULL;
 		foreach($caches as $cache)
 		{
 			@unlink($cache->id());
@@ -47,9 +76,13 @@ class cache_static extends base_object_db
 				$d = dirname($d);
 			}
 
-			if(!file_exists($cache->id()))
+			if(!file_exists($cache->id()) && !$first)
 				$cache->delete(false);
+			else
+				$first = false;
 		}
+
+		$object->set_attr('static_recreate_object', $cache);
 
 		if($object->cache_static_recreate())
 		{
@@ -76,17 +109,46 @@ class cache_static extends base_object_db
 
 		bors()->changed_save();
 
-		$cache = bors_new('cache_static', array(
-			'id' => $file,
-			'object_uri' => $object->url($object->page()),
-			'original_uri' => $object->called_url(),
-			'target_class_name' => $object->extends_class_name(),
-			'target_class_id' => $object->extends_class_id(),
-			'target_id' => $object->id(),
-			'last_compile' => time(),
-			'expire_time' => time() + ($expire_time === false ? $object->cache_static() : $expire_time),
-			'recreate' => $object->cache_static_recreate(),
-		));
+		if(!($cache = $object->attr('static_recreate_object')))
+			$cache = bors_load_ex(__CLASS__, $file, array('no_load_cache' => true));
+
+		$object_uri = $object->url($object->page());
+		$original_uri = $object->called_url();
+
+		if($cache)
+		{
+//			echo "Update $file\n";
+			$cache->set_target_class_name($object->extends_class_name(), true);
+			$cache->set_target_class_id($object->extends_class_id(), true);
+			$cache->set_target_id($object->id(), true);
+			$cache->set_last_compile(time(), true);
+			$cache->set_expire_time(time() + ($expire_time === false ? $object->cache_static() : $expire_time), true);
+			$cache->set_recreate($object->cache_static_recreate(), true);
+
+			if($object_uri)
+				$cache->set('object_uri', $object_uri, true);
+
+			if($original_uri)
+				$cache->set('original_uri', $original_uri, true);
+
+			$cache->store();
+		}
+		else
+		{
+//			echo "New $file\n";
+			$cache = bors_new('cache_static', array(
+				'id' => $file,
+				'object_uri' => $object_uri,
+				'original_uri' => $original_uri,
+				'target_class_name' => $object->extends_class_name(),
+				'target_class_id' => $object->extends_class_id(),
+				'target_id' => $object->id(),
+				'last_compile' => time(),
+				'expire_time' => time() + ($expire_time === false ? $object->cache_static() : $expire_time),
+				'recreate' => $object->cache_static_recreate(),
+				'bors_site' => BORS_SITE,
+			));
+		}
 
 		foreach(explode(' ', $object->cache_depends_on()) as $group_name)
 			if($group_name)
